Add a title to the REST response for new notes
Because the shortnotes post type does not support titles, the REST
response does not include a title. The block editor sees no title
and uses the text "(no title) is now live." when the note is
published.

We can filter in better placeholder text—"Your note"—for a REST
response made in the edit context. It's a little silly, but it
works.

Fixes #1