Do not attempt to modify content if no global post is available
When `get_the_excerpt()` is called with an explicit `$post`, that
post is passed as context to the `get_the_excerpt` filter.

Even though `wp_trim_excerpt` receives the `$post` as an argument,
it does not provide that as context when applying `the_content` to
the post content.

In this case, and probably other related cases, a fatal error would
be generated because the global `$post` data we rely on is not
available.

A future commit may also account for a second parameter passed to
`the_content`, but I'll think that through some more. :)

See #44